micalg: convert raw micalg to lowercase for map lookup
Some mail clients send micalg values in uppercase. The current
implementation doesn't allow for anything other than lowercase.

The referenced table (RFC4880 9.4) lists the algorithms in uppercase, so
it's reasonable that a client would send a micalg in something other
than lowercase.

This patch converts the raw micalg to lowercase, and modifies a test
case to have an uppercase micalg.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>